fixup: rename mutations, hooks and utils
