+++ src/Apps/Order/Routes/Shipping2/FulfillmentDetails.tsx
@@ -39,11 +39,11 @@ import { useFeatureFlag } from "System/useFeatureFlag"
 import { SavedAddressesFragmentContainer as SavedAddresses } from "Apps/Order/Routes/Shipping2/SavedAddresses2"
 import { usePrevious } from "Utils/Hooks/usePrevious"
 import { Collapse } from "Apps/Order/Components/Collapse"
-import { useShippingContext } from "Apps/Order/Routes/Shipping2/ShippingContext"
+import { useShippingContext } from "Apps/Order/Routes/Shipping2/Hooks/useShippingContext"
 import {
   FulfillmentType,
   ShippingAddressFormValues,
-} from "Apps/Order/Routes/Shipping2/shippingUtils"
+} from "Apps/Order/Routes/Shipping2/Utils/shippingUtils"
 import { RouterLink } from "System/Router/RouterLink"
 import { useOrderTracking } from "Apps/Order/Utils/useOrderTracking"
