+++ src/Apps/Order/Routes/Shipping2/SavedAddresses2.jest.tsx
@@ -14,7 +14,7 @@ jest.mock("Utils/Hooks/useMatchMedia", () => ({
   __internal__useMatchMedia: () => ({}),
 }))
 
-jest.mock("Apps/Order/Routes/Shipping2/ShippingContext", () => {
+jest.mock("Apps/Order/Routes/Shipping2/Hooks/useShippingContext", () => {
   return {
     useShippingContext: () => ({
       parsedOrderData: {
